Reorganize project files ... This might break things but hopefully not??
