+++ src/org/rrhs/asteroids/actors/Asteroid.java
@@ -1,3 +1,5 @@
+package org.rrhs.asteroids.actors;
+
 import mayflower.*;
 
 public class Asteroid extends NetworkActor
