+++ src/org/rrhs/asteroids/actors/Laser.java
@@ -1,9 +1,11 @@
-import mayflower.*;
+package org.rrhs.asteroids.actors;
+
+import mayflower.MayflowerImage;
 
 public class Laser extends NetworkActor
 {
   // TO DO:
-  // Move laser in space 
+  // Move laser in space
   // Add collsion event with asteroid
     public Laser(int id) {
         super(id, "laser");
