+++ src/org/rrhs/asteroids/actors/Ship.java
@@ -1,3 +1,5 @@
+package org.rrhs.asteroids.actors;
+
 import mayflower.*;
 
 public class Ship extends NetworkActor
