+++ src/org/rrhs/asteroids/actors/data/ActorData.java
@@ -1,3 +1,5 @@
+package org.rrhs.asteroids.actors.data;
+
 public class ActorData
 {
     public ActorData()
