+++ src/org/rrhs/asteroids/actors/data/ShipData.java
@@ -1,0 +1,5 @@
+package org.rrhs.asteroids.actors.data;
+
+public class ShipData
+{
+}
