+++ src/org/rrhs/asteroids/network/ServerWorld.java
@@ -1,3 +1,5 @@
+package org.rrhs.asteroids.network;
+
 import mayflower.*;
 
 public class ServerWorld extends World
