+++ src/org/rrhs/asteroids/util/NetworkUtils.java
@@ -1,3 +1,5 @@
+package org.rrhs.asteroids.util;
+
 public class NetworkUtils
 {
     public enum Message
