+++ src/org/rrhs/asteroids/views/EngineerView.java
@@ -1,3 +1,8 @@
+package org.rrhs.asteroids.views;
+
+import org.rrhs.asteroids.network.Client;
+import org.rrhs.asteroids.GameState;
+
 public class EngineerView extends GameView
 {
     public EngineerView(Client client, GameState state)
